# Draft: Avoid explicit calls to select.select()