@jukuisma 2024-09-24
TODO: Benchmarking. I don't foresee this being slower than `select.select()`, but rather safe than sorry.